+++ CHANGELOG.md
@@ -24,3 +24,8 @@ Update all emitted callbacks to use AsyncResource's
 
 # 1.4.6 
 bugfix/issue#217 - do not crash when no options object is passed to the constructor
+
+# 1.4.7
+issue #235 - added instanceids to add and remove association
+Fix scene APIs #264
+Fix scene values type 'list' (Partially fix #266) #267